pyproject.toml
[build-system]
requires = ["hatchling>=1.21.1"]
build-backend = "hatchling.build"

[project]
name = "titus-core"
version = "0.1.0"
description = "Core execution engines and exchange adapters for Titus trading system"
authors = [{name = "Relic Labs"}]
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "pandas>=2.2",
  "pyarrow>=15.0",
  "python-dotenv>=1.0",
  "requests>=2.32",
  "typer>=0.12",
  "rich>=13.7",
  "hyperliquid-python-sdk>=0.21.0",
  "eth-account>=0.13.0",
]

[project.scripts]
titus-live = "titus_live.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["titus_core", "titus_live"]

[tool.pytest.ini_options]
pythonpath = ["."]

